// packages/backend/src/models/insight-yaml.ts
/**
 * Copyright 2021 Expedia, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { ItemType } from '@iex/models/item-type';

export interface InsightYaml {
  name?: string;
  description?: string;
  tags?: string[];
  itemType?: ItemType;
  creation?: {
    template?: string;
    clonedFrom?: string;
  };
  metadata?: {
    publishedDate?: string;
    team?: string;
  };
}